\name{avinash-package}
\alias{avinash-package}
\alias{avinash}
\docType{package}
\title{
\packageTitle{avinash}
}
\description{
\packageDescription{avinash}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{avinash}
\packageIndices{avinash}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{avinash}

Maintainer: \packageMaintainer{avinash}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}