% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{calc.cor}
\alias{calc.cor}
\title{calc.cor calulate correlation and approximate P-value of two matrix}
\usage{
calc.cor(x, y, method = "spearman", use = "pairwise.complete.obs")
}
\arguments{
\item{x}{: matrix}

\item{y}{: matrix}

\item{method}{: a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{use}{: "an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{method}{: matrix}
}
\value{
estimate=correlation coefficient, P=P-value
}
\description{
calc.cor calulate correlation and approximate P-value of two matrix
}
