% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.deepImmune.R
\name{impute.closest.gene}
\alias{impute.closest.gene}
\title{Title}
\usage{
impute.closest.gene(common.genes, incomplete_exp_mat, ref = NULL,
  subsample = NULL)
}
\arguments{
\item{common.genes}{common genes to imputed}

\item{incomplete_exp_mat}{incomplete expression matrix sample x genes}

\item{ref}{reference expression matrix sample x genes}
}
\description{
Title
}
