% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heatmap.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Plot heatmap}
\usage{
\method{plot}{heatmap}(dat, filename, height = 10, width = 7)
}
\arguments{
\item{dat}{matrix}

\item{filename}{(optional)}

\item{height}{of pdf file}

\item{width}{of pdf file}
}
\description{
Plot heatmap
}
